fix(04): WR-04 change RadarDimension.value and ConsistencyRadar.overall_score to float
- Prevents silent truncation of float scores to int by Pydantic
- A score of 7.9 will now be preserved as 7.9 instead of truncated to 7

+++ api/models/chart.py
@@ -38,14 +38,14 @@ class RadarDimension(BaseModel):
     """Single dimension for radar chart."""
 
     name: str
-    value: int  # 0-10 score
+    value: float  # 0-10 score
 
 
 class ConsistencyRadar(BaseModel):
     """6-dimension consistency radar chart data."""
 
     dimensions: list[RadarDimension]
-    overall_score: int
+    overall_score: float
 
 
 class IssueStat(BaseModel):
